+++ lib/embedded_photo_picker.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'src/embedded_photo_picker.dart';
+export 'src/embedded_photo_picker_capabilities.dart';
 export 'src/embedded_photo_picker_controller.dart';
 export 'src/embedded_photo_picker_media.dart';
 export 'src/embedded_photo_picker_options.dart';